refactor: isolate Helm release validation
